added more templating
