+++ Cargo.toml
@@ -13,10 +13,12 @@ homepage = "https://github.com/sn99/privacy-sexy"
 license = "MIT"
 
 [dependencies]
+chrono = "0.4.24"
 clap = { version = "4.2.7", features = ["derive"] }
 regex = "1.8.1"
 serde = { version = "1.0.162", features = ["derive"] }
 serde_yaml = "0.9.21"
+toml = "0.7.4"
 
 [dev-dependencies]
 glob = "0.3.1"
